fixed an issue for deleting a file data on errors

+++ grails-app/services/metridoc/ezproxy/EzproxyService.groovy
@@ -286,7 +286,9 @@ class EzproxyService {
     void deleteDataForFile(String nameOfFileToDelete) {
         EzFileMetaData.withNewTransaction {
             def data = EzFileMetaData.findByFileName(nameOfFileToDelete)
-            EzFileMetaData.get(data.id).delete()
+            if (data) {
+                EzFileMetaData.get(data.id).delete()
+            }
             def domainClasses = grailsApplication.domainClasses
             domainClasses.each {
                 def gormRecord = it.newInstance()
